… CHANGELOG.md
… codegen/templates/module_initializer.tmpl
… codegen/templates/module_mock_initializer.tmpl
… codegen/templates/service_mock.tmpl
… codegen/templates/tchannel_client.tmpl
… codegen/templates/tchannel_endpoint.tmpl
… examples/example-gateway/build/app/demo/endpoints/abc/abc_appdemoservice_method_call_tchannel.go
… examples/example-gateway/build/app/demo/services/xyz/mock-service/mock_init.go
… examples/example-gateway/build/app/demo/services/xyz/mock-service/mock_service.go
… examples/example-gateway/build/app/demo/services/xyz/module/init.go
… examples/example-gateway/build/clients/baz/baz.go
… examples/example-gateway/build/clients/corge/corge.go
… examples/example-gateway/build/endpoints/bounce/bounce_bounce_method_bounce_tchannel.go
… examples/example-gateway/build/endpoints/tchannel/baz/baz_simpleservice_method_call_tchannel.go
… examples/example-gateway/build/endpoints/tchannel/baz/baz_simpleservice_method_echo_tchannel.go
… examples/example-gateway/build/endpoints/tchannel/echo/echo_echo_method_echo_tchannel.go
… examples/example-gateway/build/endpoints/tchannel/panic/panic_simpleservice_method_anothercall_tchannel.go
… examples/example-gateway/build/endpoints/tchannel/quux/quux_simpleservice_method_echostring_tchannel.go
… examples/example-gateway/build/services/echo-gateway/mock-service/mock_init.go
… examples/example-gateway/build/services/echo-gateway/mock-service/mock_service.go
… examples/example-gateway/build/services/echo-gateway/module/init.go
… examples/example-gateway/build/services/example-gateway/mock-service/mock_init.go
… examples/example-gateway/build/services/example-gateway/mock-service/mock_service.go
… examples/example-gateway/build/services/example-gateway/module/init.go
… examples/example-gateway/config/production.yaml
… examples/example-gateway/config/test.yaml
… examples/selective-gateway/build/endpoints/bounce/bounce_bounce_method_bounce_tchannel.go
… examples/selective-gateway/build/services/selective-gateway/mock-service/mock_init.go
… examples/selective-gateway/build/services/selective-gateway/mock-service/mock_service.go
… examples/selective-gateway/build/services/selective-gateway/module/init.go